Show the UI some love
Introduce basic UI browsing
Use SysfsVector methods instead of PrintDirContent in getHotPlugInfo()
Bump to BETA

+++ include/classes.hpp
@@ -36,7 +36,7 @@ class SysfsVector {
 
 	private:
 		vector_pair v;
-		vector <double> vf;
+		vector <unsigned long> vf;
 		string path;
 		string poss_path;
